% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gliph2rs-package.R
\docType{package}
\name{gliph2rs-package}
\alias{gliph2rs}
\alias{gliph2rs-package}
\title{gliph2rs: Fast GLIPH2 Convergence Groups for TCR Repertoires}
\description{
Groups T-cell receptor CDR3 beta sequences that are predicted to share antigen specificity, following the GLIPH2 method of Huang et al. (2020) \doi{10.1038/s41587-020-0505-4}. Group membership matches 'turboGliph' exactly, while the computation runs in parallel Rust and scales to pooled cohorts of 100 million sequences. Also tests groups for association with a donor condition, such as cytomegalovirus serostatus.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mayankgandhi13/gliph2-rs}
  \item Report bugs at \url{https://github.com/mayankgandhi13/gliph2-rs/issues}
}

}
\author{
\strong{Maintainer}: Mayank Gandhi \email{gandhi.may@northeastern.edu}

Authors:
\itemize{
  \item Mayank Gandhi \email{gandhi.may@northeastern.edu}
}

}
\keyword{internal}
